--- script.speedtester/CREDITS.txt
@@ -1,1 +1,0 @@
-Icon made by Freepik from www.flaticon.com
