+++ script.speedtester/README.md
@@ -1,0 +1,28 @@
+[![GitHub release](https://img.shields.io/github/release/add-ons/script.speedtester.svg)](https://github.com/add-ons/script.speedtester/releases)
+[![Build status](https://github.com/add-ons/script.speedtester/workflows/CI/badge.svg)](https://github.com/add-ons/script.speedtester/actions)
+[![Codecov status](https://img.shields.io/codecov/c/github/add-ons/script.speedtester/master)](https://codecov.io/gh/add-ons/script.speedtester/branch/master)
+[![License: Apache-2.0](https://img.shields.io/badge/License-Apache2-yellow.svg)](https://opensource.org/licenses/Apache-2.0)
+[![Contributors](https://img.shields.io/github/contributors/add-ons/script.speedtester.svg)](https://github.com/add-ons/script.speedtester/graphs/contributors)
+
+# Kodi Speedtester add-on
+**script.speedtester** is a [Kodi](https://kodi.tv/) add-on for testing your Internet bandwidth.
+
+
+## Installing
+This add-on ships with Kodi.
+
+In Kodi, simply search the add-ons for `Speed` and install the Speedtester add-on.
+
+
+## Releases
+### v1.1.0 (2020-09-26)
+- Move add-on to Add-ons project (@dagwieers)
+- Fix typos and structure of standard output (@dagwieers)
+
+### v1.0.1 (2019-08-06)
+- Initial release (@Dr0idGuy)
+
+
+## Thanks
+- Original add-on by Dr0idGuy
+- Icon made by Freepik from www.flaticon.com
