+++ script.speedtester/addon.xml
@@ -1,23 +1,37 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="script.speedtester"
-       name="Speed Tester"
-       version="1.0.0"
-       provider-name="Dr0idGuy">
+<addon id="script.speedtester" name="Speed Tester" version="1.1.0" provider-name="Dr0idGuy, dagwieers">
   <requires>
     <import addon="xbmc.python" version="2.26.0"/>
   </requires>
-  <extension point="xbmc.python.script"
-             library="testit.py">
-        <provides>executable</provides>
+  <extension point="xbmc.python.script" library="default.py">
+    <provides>executable</provides>
   </extension>
   <extension point="xbmc.addon.metadata">
-    <summary lang="en_GB">Speed Tester</summary>
-    <description lang="en_GB">Speed Tester powered by speedtest.net</description>
-    <license>Apache 2.0</license>
     <platform>all</platform>
+    <summary lang="en_GB">Test your Internet bandwidth</summary>
+    <description lang="en_GB">This add-on tests your Internet bandwith and reports both download and upload speeds and latency from within Kodi.
+
+Use this if you are experiencing streaming issues and you suspect your network to be the cause.</description>
+    <disclaimer lang="en_GB">Speed Tester is powered by speedtest.net</disclaimer>
+    <language>en</language>
+    <license>Apache-2.0</license>
+    <email>dag@wieers.com</email>
+    <website>https://kodi.wiki/view/Add-on:Speed_Tester</website>
+    <source>https://github.com/add-ons/script.speedtester</source>
+    <news>
+v1.1.0 (2020-09-26)
+- Move add-on to Add-ons project
+- Fix typos and structure of standard output
+
+v1.0.1 (2019-08-06)
+- Initial release
+    </news>
     <assets>
-       <icon>icon.png</icon>
-       <fanart>fanart.jpg</fanart>
+      <icon>resources/media/icon.png</icon>
+      <screenshot>resources/media/screenshot01.jpg</screenshot>
+      <screenshot>resources/media/screenshot02.jpg</screenshot>
+      <screenshot>resources/media/screenshot03.jpg</screenshot>
     </assets>
+    <reuselanguageinvoker>true</reuselanguageinvoker>
   </extension>
 </addon>
